Updated demo index.html to source everything from bower_components
Was previously getting components from a mixture of locations - CDNs,
github and the local bower_components - despite everything being available
locally in bower_components. I updated it to source everything from the local
bower_components folder

This is more consistent and ensures that the demo is using the
correct versions of all components, rather than a mixture.